--- client-dashboard/src/views/dashboard/SalesInvoice.js
@@ -1,253 +1,0 @@
-import * as React from 'react';
-import { useState, useEffect } from 'react';
-import PropTypes from 'prop-types';
-import { useTheme } from '@mui/material/styles';
-import {
-  Typography,
-  TableHead,
-  Chip,
-  Box,
-  Table,
-  TableBody,
-  TableCell,
-  TablePagination,
-  TableRow,
-  TableFooter,
-  IconButton,
-  Paper,
-  TableContainer,
-} from '@mui/material';
-
-import FirstPageIcon from '@mui/icons-material/FirstPage';
-import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
-import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
-import LastPageIcon from '@mui/icons-material/LastPage';
-
-import Breadcrumb from '../../layouts/full/shared/breadcrumb/Breadcrumb';
-import PageContainer from '../../components/container/PageContainer';
-
-import ParentCard from '../../components/shared/ParentCard';
-import { Stack } from '@mui/system';
-
-function TablePaginationActions(props) {
-  const theme = useTheme();
-  const { count, page, rowsPerPage, onPageChange } = props;
-
-  const handleFirstPageButtonClick = (event) => {
-    onPageChange(event, 0);
-  };
-
-  const handleBackButtonClick = (event) => {
-    onPageChange(event, page - 1);
-  };
-
-  const handleNextButtonClick = (event) => {
-    onPageChange(event, page + 1);
-  };
-
-  const handleLastPageButtonClick = (event) => {
-    onPageChange(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-  };
-
-  return (
-    <Box sx={{ flexShrink: 0, ml: 2.5 }}>
-      <IconButton
-        onClick={handleFirstPageButtonClick}
-        disabled={page === 0}
-        aria-label="first page"
-      >
-        {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
-      </IconButton>
-      <IconButton onClick={handleBackButtonClick} disabled={page === 0} aria-label="previous page">
-        {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
-      </IconButton>
-      <IconButton
-        onClick={handleNextButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="next page"
-      >
-        {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
-      </IconButton>
-      <IconButton
-        onClick={handleLastPageButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="last page"
-      >
-        {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
-      </IconButton>
-    </Box>
-  );
-}
-
-TablePaginationActions.propTypes = {
-  count: PropTypes.number.isRequired,
-  onPageChange: PropTypes.func.isRequired,
-  page: PropTypes.number.isRequired,
-  rowsPerPage: PropTypes.number.isRequired,
-};
-
-// const rows = [
-//   {
-//     invoiceno: 'MPSS23240001',
-//     invoicedate: '01-04-2023',
-//     invoiceamt: '2,348.00',
-//     invoiceduedate: '11-04-2023',
-//     status: 'Overdue',
-//   },
-
-// ].sort((a, b) => (a.calories < b.calories ? -1 : 1));
-
-const BCrumb = [
-  {
-    to: '/dashboard',
-    title: 'Home',
-  },
-  {
-    title: 'Sales Invoice',
-  },
-];
-
-const SalesInvoice = () => {
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(5);
-  const [rows, setRow] = useState([].sort((a, b) => (a.calories < b.calories ? -1 : 1)));
-
-  useEffect(() => {
-
-    fetch("https://erpnoveloffice.in/api/method/sales-invoice-api?message=overdue")
-      .then((res) => {
-        if (res.ok === false) {
-          throw Error("Searching data not found in this API")
-        }
-        return res.json()
-      })
-      .then((data) => { setRow(data.message) })
-      .catch((err) => { console.log(err)})
-
-  },[]);
-
-  console.log(rows);
-
-  // Avoid a layout jump when reaching the last page with empty rows.
-  const emptyRows = page > 0 ? Math.max(0, (1 + page) * rowsPerPage - rows.length) : 0;
-
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(parseInt(event.target.value, 10));
-    setPage(0);
-  };
-
-  return (
-    <PageContainer title="Sales Invoice" description="this is Sales Invoice page">
-      {/* breadcrumb */}
-      <Breadcrumb title="Sales Invoice" items={BCrumb} />
-      {/* end breadcrumb */}
-      <ParentCard title="Overdues">
-        <Paper variant="outlined">
-          <TableContainer>
-            
-            <Table
-              aria-label="custom sales invoice"
-              sx={{
-                whiteSpace: 'nowrap',
-              }}
-            >
-              <TableHead>
-                <TableRow>
-                  <TableCell>
-                    <Typography variant="h6">Invoice No.</Typography>
-                  </TableCell>
-                  <TableCell>
-                    <Typography variant="h6">Invoice Date</Typography>
-                  </TableCell>
-                  <TableCell>
-                    <Typography variant="h6">Invoice Amount</Typography>
-                  </TableCell>
-
-                  <TableCell>
-                    <Typography variant="h6">Due Date</Typography>
-                  </TableCell>
-                  <TableCell>
-                    <Typography variant="h6">Status</Typography>
-                  </TableCell>
-                </TableRow>
-              </TableHead>
-
-              <TableBody>
-                {(rowsPerPage > 0
-                  ? rows.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                  : rows
-                ).map((rows, index) => (
-                  <TableRow key={index}>
-                    <TableCell>
-                      <Typography variant="h6">{rows.name}</Typography>
-                    </TableCell>
-
-                    <TableCell>
-                      <Typography color="textSecondary" variant="h6" fontWeight="400">
-                        {rows.lr_date}
-                      </Typography>
-                    </TableCell>
-
-                    <TableCell>
-                      <Typography color="textSecondary" variant="h6" fontWeight="400">
-                        ₹ {rows.outstanding_amount}
-                      </Typography>
-                    </TableCell>
-
-                    <TableCell>
-                      <Typography variant="h6">{rows.due_date}</Typography>
-                    </TableCell>
-                    <TableCell>
-                      <Chip color={rows.status === 'Overdue' ? 'error' : 'secondary'}
-                        sx={{
-                          borderRadius: '6px',
-                        }}
-                        size="small"
-                        label={rows.status}
-                      />
-                    </TableCell>
-                  </TableRow>
-                ))}
-
-                {emptyRows > 0 && (
-                  <TableRow style={{ height: 53 * emptyRows }}>
-                    <TableCell colSpan={6} />
-                  </TableRow>
-                )}
-              </TableBody>
-
-              <TableFooter>
-                <TableRow>
-                  <TablePagination
-                    rowsPerPageOptions={[5, 10, 25, { label: 'All', value: -1 }]}
-                    colSpan={6}
-                    count={rows.length}
-                    rowsPerPage={rowsPerPage}
-                    page={page}
-                    SelectProps={{
-                      inputprops: {
-                        'aria-label': 'rows per page',
-                      },
-                      native: true,
-                    }}
-                    onPageChange={handleChangePage}
-                    onRowsPerPageChange={handleChangeRowsPerPage}
-                    ActionsComponent={TablePaginationActions}
-                  />
-                </TableRow>
-              </TableFooter>
-
-            </Table>
-
-          </TableContainer>
-        </Paper>
-      </ParentCard>
-    </PageContainer>
-  );
-};
-
-export default SalesInvoice;
